requirements.txt
streamlit
sqlparse
sqlglot
pandas
python-dotenv
tiktoken
openai
psycopg2-binary